requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ajax-selects==2.2.0
django-markdownx==4.0.2
importlib-metadata==6.1.0
Markdown==3.4.3
Pillow==9.5.0
sqlparse==0.4.3
zipp==3.15.0
